Parse and cache mappings upfront
The original purpose of this was to combine all mappings into a global list, in order to include friendly names added later. But as far as I know, there is no link between a friendly name and a generic persistent name. Either way, the code is a bit cleaner this way. We can easily defer LoadLocalMappings later if doing it upfront is a problem for whatever reason.